Refactor database models and dependencies: remove redundant triggers, update foreign key relationships, centralize `create_all` into `__init__.py`, enhance logging, and streamline imports and DDL execution logic.
