+++ src/scripts/reload_ddl.py
@@ -1,5 +1,15 @@
-from src.database.base_model import create_all
+import logging
+
+from src.config import settings
+from src.database import create_all
 from src.database.db_connector import create_connection, ConnectionCM
+
+logging.basicConfig(
+    level=settings.logging.level,
+    format=settings.logging.format,
+    datefmt=settings.logging.datefmt,
+    force=True,  # overrides existing logging configuration (useful for repeated runs)
+)
 
 if __name__ == '__main__':
     db_connection = create_connection()
